+++ doc/version_history.rst
@@ -11,6 +11,7 @@ v0.6.7
 
 * Update the code that handles the authentication response.
 * Make sure that the CSC goes to FAULT state in case of an error in the RestAuthorizeHandler.
+* Handle the POST response code 201 correctly.
 
 Requirements: